requirements.txt
dbus-next
sdbus-networkmanager
systemd-python
sdbus
